Moved points out of an excel file into a python file
